third_party/WebKit/LayoutTests/platform/win/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_redirects-expected.txt

CONSOLE WARNING: line 4: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
This is a testharness.js-based test.
PASS http://127.0.0.1:8000/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=css is expected to have initiatorType link 
PASS http://127.0.0.1:8000/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=iframe is expected to have initiatorType iframe 
PASS http://127.0.0.1:8000/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=iframe%26id=xhr is expected to have initiatorType xmlhttprequest 
PASS http://127.0.0.1:8000/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=image is expected to have initiatorType img 
PASS http://127.0.0.1:8000/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=script is expected to have initiatorType script 
Harness: the test ran to completion.
